ui/requirements.txt
textual
websocket-client
pytest